+++ packages/fast-tooling-react/src/css-editor/editor.style.ts
@@ -1,6 +1,6 @@
 import { ComponentStyles } from "@microsoft/fast-jss-manager-react";
 import { CSSPositionClassNameContract } from "./position/position.style";
-import { background300, foreground300 } from "../style";
+import { foreground300, neutralLayerL4 } from "../style";
 
 export interface CSSEditorClassNameContract extends CSSPositionClassNameContract {
     cssEditor?: string;
@@ -8,7 +8,7 @@ export interface CSSEditorClassNameContract extends CSSPositionClassNameContract
 
 const styles: ComponentStyles<CSSEditorClassNameContract, {}> = {
     cssEditor: {
-        background: background300,
+        background: neutralLayerL4,
         color: foreground300,
         height: "100%",
         padding: "0 10px",