+++ packages/fast-tooling-react/src/navigation/navigation-tree-item.props.ts
@@ -16,7 +16,7 @@ export interface NavigationTreeItemProps
     /**
      * The class
      */
-    className: string;
+    className: (dragging: boolean) => string;
 
     /**
      * The class for the content
@@ -42,21 +42,6 @@ export interface NavigationTreeItemProps
      * The keyUp handler
      */
     handleKeyUp: React.KeyboardEventHandler<HTMLElement>;
-
-    /**
-     * The level, used for the tree items aria-level
-     */
-    level: number;
-
-    /**
-     * The length of the navigation, used for the tree items aria-setsize
-     */
-    navigationLength: number;
-
-    /**
-     * The position of the tree item in the navigation, used for the tree items aria-posinset
-     */
-    positionInNavigation: number;
 
     /**
      * The text used for the tree item