+++ packages/fast-color-explorer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-color-explorer@1.1.0...@microsoft/fast-color-explorer@1.1.1) (2019-07-01)
+
+**Note:** Version bump only for package @microsoft/fast-color-explorer
+
+
+
+
+
 # [1.1.0](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-color-explorer@1.0.11...@microsoft/fast-color-explorer@1.1.0) (2019-06-25)